# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "API scrapes news from various sources and provides a unified interface for users to view and manage their preferences and newsfeed also filters news based on user preferences",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/user/preferences": {
            "get": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Get user preferences",
                "description": "This endpoint fetches the user's preferences.",
                "operationId": "9e1670cec2ff4ca3cd12dd3701fbc8cd",
                "responses": {
                    "200": {
                        "description": "User preferences fetched successfully"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Update user preferences",
                "description": "This endpoint updates the user's preferences.",
                "operationId": "23aa1b0cfd017c426b22be7dbb542357",
                "requestBody": {
                    "description": "User preference data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "sources": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User preferences updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User preferences updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 400
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Invalid request"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "example": 500
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Something went wrong"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/newsfeed": {
            "get": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Get newsfeed",
                "description": "Fetches newsfeed based on user preferences",
                "operationId": "dd99a96754d3f541740e97ede1200d36",
                "responses": {
                    "200": {
                        "description": "Newsfeed fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "collection"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}